html/semantics/scripting-1/the-script-element/scriptexecute-events/020.html

<!DOCTYPE html>
<html>
  <title>Cancelling onbeforescriptexecute for external script</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <div id="log"></div>
  <script>
    targets = [];
    var t = async_test();
    document.addEventListener("beforescriptexecute",
      function x(e) {
       targets.push(e.target);
       e.preventDefault();
       document.removeEventListener("beforescriptexecute", x);
    }, false);
 </script>
 <script src="include-2.js"></script>
 <script>
   t.step(function(){assert_array_equals(targets, [document.getElementsByTagName("script")[3]])});
   t.done();
 </script>
